Move getProcessForPort to react-dev-utils
- also allowed for breakdown of commands into helper methods
